Planner: extract schema module to planner/schema.rs; add column_name_map
